Module Cduce.Tvar

module TVar : sig ... end

module TVH : Stdlib.Hashtbl.S with type key = TVar.t

module TVarSet : sig ... end

module Subst : sig ... end

val vars : Base.typ -> TVarSet.t

val vars_mono : Base.typ -> TVarSet.t

val vars_poly : Base.typ -> TVarSet.t

val vars_infer : Base.typ -> TVarSet.t

val top_vars : Base.typ -> TVarSet.t

val vars_with_polarity : Base.typ -> (TVar.t * [ `Both | `Neg | `Pos ]) list

val check_var : Base.typ -> [ `Not_var | `Pos of TVar.t | `Neg of TVar.t ]

val is_mono_typ : Base.typ -> bool

val is_novar_typ : Base.typ -> bool

val refresh : TVarSet.t -> Subst.t

val generalize : TVarSet.t -> Subst.t

val monomorphize : TVarSet.t -> Subst.t

val pp_typ_short : Stdlib.Format.formatter -> Base.typ -> unit

val string_of_type_short : Base.typ -> string

module Raw : sig ... end

This module contains raw Cduce functions. In particular, they make no distinction between between polymorphic type variables and monomorphic type variables. Functions defined outside of this module should be preferred.

val clean_type : pos:Base.typ -> neg:Base.typ -> Base.typ -> Base.typ

val clean_type_subst : pos:Base.typ -> neg:Base.typ -> Base.typ -> Subst.t

val test_tallying : (Base.typ * Base.typ) list -> bool

val tallying : (Base.typ * Base.typ) list -> Subst.t list

val tallying_infer : (Base.typ * Base.typ) list -> Subst.t list

val factorize : (TVarSet.t * TVarSet.t) -> Base.typ -> Base.typ * Base.typ