Module Tvar.Raw

This module contains raw Cduce functions. In particular, they make no distinction between between polymorphic type variables and monomorphic type variables. Functions defined outside of this module should be preferred.

val clean_type : pos:Base.typ -> neg:Base.typ -> TVarSet.t -> Base.typ -> Base.typ

clean_type p n mono t substitutes in t all type variables not in mono and only occurring positively by p, and all type variables not in mono and only occurring negatively by n

val rectype : Base.typ -> TVar.t -> Base.typ

rectype t u returns the type corresponding to the equation u=t

val tallying : TVarSet.t -> (Base.typ * Base.typ) list -> Subst.t list

val tallying_with_prio : TVar.t list -> TVarSet.t -> (Base.typ * Base.typ) list -> Subst.t list

val test_tallying : TVarSet.t -> (Base.typ * Base.typ) list -> bool