Module Sstt.LabelMap

type key = Label.t

The type of keys

type value = Ty.F.t

The type of values

type t

Maps of hashable and comparable keys and values which support hash in constant time.

module Set = VDescr.Descr.Records.Atom.LabelMap.Set

Represent sets of keys

val add : key -> value -> t -> t

Add a binding to a map. If a binding already exists, the old value is replaced.

val dom : t -> Set.t

The domain of the map (in constant time).

val bindings : t -> (key * value) list

Return the list of bindings.

val compare : t -> t -> int

Total ordering between maps.

val equal : t -> t -> bool

Equality between two maps.

val hash : t -> int

Hash value of a map (in constant time).

val empty : t

The empty map.

val exists : (key -> value -> bool) -> t -> bool

Test whether a binding verifies a predicate.

val filter : (key -> value -> bool) -> t -> t

Return the map of bindings which verify a predicate.

val find_opt : key -> t -> value option

Find a value in the map. Runs in linear time in the number of bindings.

val map : (value -> value) -> t -> t

Return a map with the same keys and updated values

val filter_map : (key -> value -> value option) -> t -> t

val of_list : (key * value) list -> t

Build from a list of bindings

val singleton : key -> value -> t

Return the map with one binding.

val to_list : t -> (key * value) list

An alias for bindings.

val merge : (key -> value option -> value option -> value option) -> t -> t -> t

Merge two maps. For each key the merge function is called. The optional value given as first (resp. second) arguments determines the presence of the binding in the first (resp. second) map. If the merge function returns None, no binding is created for that key.

val values_for_domain : Set.t -> value -> t -> value list

values_for_domain dom def map returns a map whose domain is dom and whose value is:

taken from map if the binding exists there

otherwise def

val values : t -> value list

Return the values of a map.

val constant : Set.t -> value -> t

Return a constant map with from the given domain.

val fold : ('a -> key -> value -> 'a) -> 'a -> t -> 'a

Folding over the bindings of the map.

val is_singleton_opt : t -> (key * value) option

If the map is a singleton, returns its single binding as Some (k, v). Otherwise returns None.

val combine : Set.t -> value list -> t

Create a map from the given domain and values, taken in the order of the list.

raises Invalid_arugment
if the domain and list of values do not have the same length.