Module Sstt.LabelSet

Represent sets of keys

type elt = Label.t

The type of elements in the set.

type t

Sets of comparable and hashable elements which support hash in constant time. We call the view of a set is the list (x0,h0) :: (x1, h1) :: … :: [] where the xi are the ordered elements of the set and hi is the has computed between the has of element xi and the tail of the list starting with element xi+1.

val compare : t -> t -> int

Total order between two sets implemented as lexicographic order over the views.

val equal : t -> t -> bool

Equal is structural equality between the views.

val hash : t -> int

Compute the hash of the set, in constant time.

val empty : t

The empty set.

val is_empty : t -> bool

val subset : t -> t -> bool

val elements : t -> elt list

Return the elements of the set as a sorted list.

val filter : (elt -> bool) -> t -> t

Filtering, filter f s is the set of elements of s for which f returns true.

val mem : elt -> t -> bool

Membership test.

val of_list : elt list -> t

Convert from a list of elements which may contain duplicates w.r.t to the comparison function. (which element is kept in this case is unspecified).

val to_list : t -> elt list

An alias for elements.

val union : t -> t -> t

The union of two sets. If an element compares equal in both set, the element of the first set is kept.

val inter : t -> t -> t

The intersection of two sets. If an element compares equal in both set, the element of the first set is kept.

val diff : t -> t -> t

The difference of two sets.

val cardinal : t -> int

Return the number of elements in the set.