Module Op.Arrows

Operations on arrow types.

type t = Arrows.t

val dom : t -> Ty.t

dom t returns the domain of the arrow component t.

val apply : t -> Ty.t -> Ty.t

apply t arg returns the type resulting from the application of an argument of type arg to a function t. An argument not in the domain will yield the resulting type any.

val worra : t -> Ty.t -> Ty.t

worra t res returns the type that must necessarily have an argument applied to the function t for the result to have type res (assuming the application did not diverge). For instance, for an arrow type t\equiv (\texttt{int}\rightarrow\texttt{bool})\cap(\texttt{string}\rightarrow\texttt{int}), \texttt{worra} t \texttt{bool} returns \texttt{int} (worra is arrow in reverse).