Module Op.Records

Operations on record types, ignores row variables.

module Atom : sig ... end

type t = Records.t

type atom = Atom.t

val of_atom : atom -> t

of_atom t returns the record component of an atom.

val as_union : t -> atom list

as_union t over-approximates t as an union of non-empty atoms.

val of_union : atom list -> t

of_union atoms returns the record component composed of the union atoms.

val approx : t -> atom

approx t over-approximates t as a non-empty atom.

raises EmptyAtom
if t is empty.

val proj : Label.t -> t -> Ty.F.OTy.t

proj l t returns the (possibly absent) type resulting from the projection on the label l of t.

val merge : atom -> atom -> t

merge t1 t2 returns the atom resulting from the merging of t1 and t2 (non-absent fields in t2 override those in t1).

val remove : atom -> Label.t -> t

remove t l returns the atom obtained by making the field l absent in t.