Module Op.TagComp

Operations on tag types.

type t = TagComp.t

type atom = VDescr.Descr.Tags.Comp.Atom.t

val is_identity : t -> bool

is_identity t returns true if and only if the component t is a tag component whose underlying interpretation is isomorphic to the identity, that is, if it is monotonic and \cap-\cup-preserving.

val preserves_cap : t -> bool

preserves_cap t returns true if and only if the component t is a tag component whose underlying interpretation preserves intersections.

val preserves_cup : t -> bool

preserves_cup t returns true if and only if the component t is a tag component whose underlying interpretation preserves unions.

val as_atom : t -> atom

as_atom t expresses t as an atom. Raises: Invalid_argument if the tag component does not satisfy is_identity.

val as_union : t -> atom list

as_union t expresses t as a union of atoms. Raises: Invalid_argument if the tag component does not satisfy preserves_cap.