Module Sstt.Prec

Precedence of operators and associativity

type unop =

| Neg

Unary operators

type funop =

| FNeg

Unary field operators

type binop =

| Diff

| Arrow

Binary operators

type fbinop =

| FDiff

Binary field operators

type varop =

| Tuple

| Cup

| Cap

Variadic operators

type fvarop =

| FCup

| FCap

Variadic field operators

type assoc =

| Left

| Right

| NoAssoc

Associativity

val fs : ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6 -> ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6

val varop_info : varop -> ('a, 'b, 'c, 'd, 'd, 'a) Stdlib.format6 * int * assoc

Returns the separator, the priority (as an integer) and the associativity of a variadic operator.

val fvarop_info : fvarop -> ('a, 'b, 'c, 'd, 'd, 'a) Stdlib.format6 * int * assoc

val binop_info : binop -> ('a, 'b, 'c, 'd, 'd, 'a) Stdlib.format6 * int * assoc

Returns the separator, the priority (as an integer) and the associativity of a binary operator.

val fbinop_info : fbinop -> ('a, 'b, 'c, 'd, 'd, 'a) Stdlib.format6 * int * assoc

val unop_info : unop -> ('a, 'b, 'c, 'd, 'd, 'a) Stdlib.format6 * int * assoc

Returns the separator, the priority (as an integer) and the associativity of a unary operator.

val funop_info : funop -> ('a, 'b, 'c, 'd, 'd, 'a) Stdlib.format6 * int * assoc

val max_prec : int

Maximum priority

val min_prec : int

Minimum priority

val need_parentheses : int -> assoc -> (('a, 'b, 'c, 'd) Stdlib.format4 * int * assoc) -> bool

need_parentheses lvl assoc info returns true if the operator described by info needs parentheses for the current printing level lvl and the current associativity assoc.

val fprintf : int -> assoc -> (('a, 'b, 'c, 'd) Stdlib.format4 * int * assoc) -> Stdlib.Format.formatter -> ('e, Stdlib.Format.formatter, unit, unit, unit, unit) Stdlib.format6 -> 'e

fprintf lvl assoc info fmt f … works as Format.fprintf but will add parentheses around the formatted output if required by lvl assoc and info.

val print_seq : (Stdlib.Format.formatter -> 'a -> unit) -> (unit, Stdlib.Format.formatter, unit) Stdlib.format -> Stdlib.Format.formatter -> 'a list -> unit

val print_nary_op : string -> (int -> assoc -> Stdlib.Format.formatter -> 'a -> unit) -> int -> assoc -> varop -> Stdlib.Format.formatter -> 'a list -> unit

val print_cup : (int -> assoc -> Stdlib.Format.formatter -> 'a -> unit) -> int -> assoc -> Stdlib.Format.formatter -> 'a list -> unit

val print_cap : (int -> assoc -> Stdlib.Format.formatter -> 'a -> unit) -> int -> assoc -> Stdlib.Format.formatter -> 'a list -> unit

val print_neg : (int -> assoc -> Stdlib.Format.formatter -> 'a -> unit) -> int -> assoc -> Stdlib.Format.formatter -> 'a -> unit

val print_lit : (int -> assoc -> Stdlib.Format.formatter -> 'a -> unit) -> int -> assoc -> Stdlib.Format.formatter -> (bool * 'a) -> unit

val print_line : any:string -> (int -> assoc -> Stdlib.Format.formatter -> 'a -> unit) -> int -> assoc -> Stdlib.Format.formatter -> ('a list * 'a list) -> unit

val print_non_empty_dnf : any:string -> (int -> assoc -> Stdlib.Format.formatter -> 'a -> unit) -> int -> assoc -> Stdlib.Format.formatter -> ('a list * 'a list) list -> unit

val print_dnf : empty:string -> any:string -> (int -> assoc -> Stdlib.Format.formatter -> 'a -> unit) -> int -> assoc -> Stdlib.Format.formatter -> ('a list * 'a list) list -> unit