Module Sstt.Records

Record components.

type t

Record types is a union of intersection of positive and negative records:

        \bigcup_{i=1\ldots m} \bigcap_{j=1 \ldots p} \texttt{\{}l_{j}{^1}:o_{j}^1,\ldots,l_{j}^k:o_{j}^{k} ~~ b_j\texttt{\}} \cap
        \bigcap_{j=1 \ldots n} \lnot(\{l_{j}{^1}:o_{j}^1,\ldots,l_{j}^l:o_{j}^{l} ~~ b_j\})
      

with b_j \in \{\texttt{..},~ \epsilon\}.

module FTy = Ty.F

val any : t

The top element of t (the set of all values of type t).

val empty : t

The bottom element of t (the empty set of values of type t).

val compare : t -> t -> int

Comparison working on the internal representation of t.

val equal : t -> t -> bool

Equality, equal a b is equivalent to compare a b = 0.

val hash : t -> int

Hashing, hash is consistent with equality, that is equal a b implies hash a = hash b.

Basics

The minimal signature of a component.

module Atom = VDescr.Descr.Records.Atom

Explicit DNF, construction and extraction

val dnf : t -> (Atom.t list * Atom.t list) list

dnf t returns a disjunctive normal form of t.

val of_dnf : (Atom.t list * Atom.t list) list -> t

of_dnf d builds a component from a disjunctive normal form d.

val mk : Atom.t -> t

Creates a component from an atom.

val map_nodes : (Ty.t -> Ty.t) -> t -> t

map_nodes f t replaces every node n in t by the node f n.

val map : (Atom.t -> Atom.t) -> t -> t

map f t replaces every atom a in t by f a.

val dnf_line_to_types : (Atom.t list * Atom.t list) -> (FTy.t * FTy.t list) * (FTy.t * FTy.t list) list

dnf_line_to_types (ps, ns) converts a line of a Record DNF to a tuple (p', ns'), where p' is a pair composed of the tail and field types of the positive side, and ns is a list pair composed of the tail and field types of the negative side.

Records have a condensed representation where each record atom requires the existence of extra labels. For instance, given the (closed) record

      \texttt{\{} x:t, y:s \texttt{\}}
      

its negation can be written as a union of positive condensed atoms:

      \texttt{\{} x:\lnot t, \texttt{.. \}}\cup
      \texttt{\{} y:\lnot s, \texttt{.. \}}\cup
      \texttt{\{} x:t, y:s, \overline{\{x, y\}}~\texttt{.. \}}
      

meaning any record which has label x associated to a type that is not t, or any record which has label y associated to a type that is not s, or any record which has both labels associated to their original type but which also has an extra label that is neither x nor y.

Condensed DNF

module Atom' = VDescr.Descr.Records.Atom'

The abstract module representing condensed atoms.

val dnf' : t -> Atom'.t list

dnf' t returns the condensed DNF of t.

val of_dnf' : Atom'.t list -> t

of_dnf' d builds a component from an condensed DNF d.