Module Ty.F

type t

module OTy : sig ... end

Optional types are subsets of 𝟙\cup\bot. They are used for the type of record fields, to denote the fact that a field may be absent.

val any : t

The top element of t (the set of all values of type t).

val empty : t

The bottom element of t (the empty set of values of type t).

val compare : t -> t -> int

Comparison working on the internal representation of t.

val equal : t -> t -> bool

Equality, equal a b is equivalent to compare a b = 0.

val hash : t -> int

Hashing, hash is consistent with equality, that is equal a b implies hash a = hash b.

Constructors

val mk_var : RowVar.t -> t

mk_var v builds a full descriptor from a given variable.

val mk_descr : OTy.t -> t

mk_descr d creates a full descriptor from the monomorphic descriptor d.

val get_descr : t -> OTy.t

get_descr t extracts a monomorphic descriptor from t, which describes t by ignoring its top-level variables.

val get_vars : t -> Stdlib__Set.Make(Sstt__.Base.RowVar).t

get_vars t extracts the top-level variables in t.

Set-theoretic

val cap : t -> t -> t

The intersection of two types, t_1 \cap t_2.

val cup : t -> t -> t

The union of two types, t_1 \cup t_2.

val diff : t -> t -> t

The difference of two types, t_1 \setminus t_2.

val neg : t -> t

The negation of a type, \lnot t.

val conj : t list -> t

conj l is the intersection of all the types in l. It returns any if l is the empty list.

val disj : t list -> t

disj l is the union of all the types in l. It returns empty if l is the empty list.

Explicit Disjunctive Normal Form

val dnf : t -> (RowVar.t list * RowVar.t list * OTy.t) list

Return an explicit DNF representation from a full descriptor.

val of_dnf : (RowVar.t list * RowVar.t list * OTy.t) list -> t

Builds a full descriptor from a DNF.

Misc. operations

val map : (OTy.t -> OTy.t) -> t -> t

map f t replaces every descriptor d in t by the descriptor f d.

val map_nodes : (Ty.t -> Ty.t) -> t -> t

map_nodes f t replaces every node n in t by the node f n.

val strengthen : (t * t) Stdlib__Map.Make(Sstt__.Base.RowVar).t -> t -> t

strengthen b t strengthens t by substituting its (top-level) type variables according to the bounds in b. For instance, if the variable v is associated to (int, int|bool), its positive occurrences will be substituted by v & (int|bool) and its negative occurrences by v | int.

val weaken : (t * t) Stdlib__Map.Make(Sstt__.Base.RowVar).t -> t -> t

weaken b t weakens t by substituting its (top-level) type variables according to the bounds in b. For instance, if the variable v is associated to (int, int|bool), its positive occurrences will be substituted by v | int and its negative occurrences by v & (int|bool).

val compare' : (Ty.t -> Ty.t -> int) -> t -> t -> int

val equal' : (Ty.t -> Ty.t -> bool) -> t -> t -> bool

val hash' : (Ty.t -> int) -> t -> int