Module Records.Atom

module LabelMap : sig ... end

type t = {

bindings : LabelMap.t;
(*
mapping from labels to field types
*)

tail : F.t;
(*
tail of the record
*)

}

A single record type.

val dom : t -> LabelMap.Set.t

dom t returns the set of explicit labels in t. Note that this does not mean that labels in dom t are present in the record values captured by t: even if a binding is present in t, it could be associated with a possibly absent type.

val find : Label.t -> t -> F.t

find l t returns the type associated with the label l in t, even if t does not have an explicit binding for l.

val default : t -> F.t option -> F.t

default t ofty returns fty if ofty is Some fty, otherwise returns the tail type for t.

val to_tuple : LabelMap.Set.t -> t -> F.t list

to_tuple lst r returns the list of fields associated with each label of lst in r.

val compare : t -> t -> int

Comparison working on the internal representation of t.

val equal : t -> t -> bool

Equality, equal a b is equivalent to compare a b = 0.

val hash : t -> int

Hashing, hash is consistent with equality, that is equal a b implies hash a = hash b.

val map_nodes : (Ty.t -> Ty.t) -> t -> t

map_nodes f r applies f to all nodes in r.bindings.

val substitute : t RowVarMap.t -> t -> t

substitute s r applies the substitution s to r.

val vars_toplevel : t -> RowVarSet.t

vars_toplevel s r returns the top-level row variables in r.