Module Comp.Atom

type t = Ty.t list

val compare : t -> t -> int

Comparison working on the internal representation of t.

val equal : t -> t -> bool

Equality, equal a b is equivalent to compare a b = 0.

val hash : t -> int

Hashing, hash is consistent with equality, that is equal a b implies hash a = hash b.

val map_nodes : (Ty.t -> Ty.t) -> t -> t