Module Records.Atom'

The abstract module representing condensed atoms.

type t = {

bindings : Atom.LabelMap.t;

tail : Ty.F.t;

exists : (Atom.LabelMap.Set.t * Ty.F.t) list;

}

A compact representation for record types. The bindings and tail fields have the same meaning as in Records.Atom.t.

val dom : t -> Atom.LabelMap.Set.t

dom t returns the set of explicit labels in t. Note that this does not mean that labels in dom t are present in the record values captured by t: even if a binding is present in t, it could be associated with a possibly absent type.

val find : Label.t -> t -> Ty.F.t

find l t returns the type associated with the label l in t, even if t does not have an explicit binding for l.

val default : t -> Ty.F.t option -> Ty.F.t

default t ofty returns fty if ofty is Some fty, otherwise returns the tail type for t.

val compare : t -> t -> int

Comparison working on the internal representation of t.

val equal : t -> t -> bool

Equality, equal a b is equivalent to compare a b = 0.

val hash : t -> int

Hashing, hash is consistent with equality, that is equal a b implies hash a = hash b.