Module Tuples.Comp

Component for a tuple of a particular fixed arity.

type t

The component of n-tuples is a union of intersections of positive and negative n-tuples.

An atom for a tuple of arity n si simply an (orederd) list of n nodes.

val compare : t -> t -> int

Comparison working on the internal representation of t.

val equal : t -> t -> bool

Equality, equal a b is equivalent to compare a b = 0.

val hash : t -> int

Hashing, hash is consistent with equality, that is equal a b implies hash a = hash b.

Basics

The minimal signature of a component.

module Atom : sig ... end

Explicit DNF, construction and extraction

val dnf : t -> (Atom.t list * Atom.t list) list

dnf t returns a disjunctive normal form of t.

val mk : Atom.t -> t

Creates a component from an atom.

val map_nodes : (Ty.t -> Ty.t) -> t -> t

map_nodes f t replaces every node n in t by the node f n.

val map : (Atom.t -> Atom.t) -> t -> t

map f t replaces every atom a in t by f a.

Operations for indexed components

Operations on components that are indexed by a type, such as tuples (indexed by integers) or tagged types (indexed by a symbolic tag).

val any : int -> t

The top element for the given index.

val empty : int -> t

The bottom element for the given index.

val of_dnf : int -> (Atom.t list * Atom.t list) list -> t

of_dnf idx d builds a component for the given index and DNF.

val len : t -> int

len t returns the common arity of all tuple types in t.

Condensed DNF

Tuples commute with intersection, so a union of intersection of tuples can be compactly represented as a union of products.

Operations on optimzed indexed components

val dnf' : t -> Atom.t list

dnf' t returns the condensed DNF of t.

val of_dnf' : int -> Atom.t list -> t

of_dnf' idx d builds a component for the given index and condensed DNF.